@@ -30,11 +30,10 @@
 
 #include "opencv2/core/core.hpp"
 
-#include "boost/optional.hpp"
-
 #include <array>
 #include <vector>
 #include <memory>
+#include <optional>
 
 namespace eos {
 	namespace render {
@@ -125,7 +124,7 @@ namespace eos {
  * @param[in] enable_far_clipping Whether vertices should be clipped against the far plane.
  * @return A pair with the colourbuffer as its first element and the depthbuffer as the second element.
  */
-inline std::pair<core::Image4u, core::Image1d> render(core::Mesh mesh, glm::tmat4x4<float> model_view_matrix, glm::tmat4x4<float> projection_matrix, int viewport_width, int viewport_height, const boost::optional<Texture>& texture = boost::none, bool enable_backface_culling = false, bool enable_near_clipping = true, bool enable_far_clipping = true)
+inline std::pair<core::Image4u, core::Image1d> render(core::Mesh mesh, glm::tmat4x4<float> model_view_matrix, glm::tmat4x4<float> projection_matrix, int viewport_width, int viewport_height, const std::optional<Texture>& texture = std::nullopt, bool enable_backface_culling = false, bool enable_near_clipping = true, bool enable_far_clipping = true)
 {
 	// Some internal documentation / old todos or notes:
 	// maybe change and pass depthBuffer as an optional arg (&?), because usually we never need it outside the renderer. Or maybe even a getDepthBuffer().
@@ -201,7 +200,7 @@ inline std::pair<core::Image4u, core::Image1d> render(core::Mesh mesh, glm::tmat
 		// all vertices are visible - pass the whole triangle to the rasterizer. = All bits of all 3 triangles are 0.
 		if ((visibility_bits[0] | visibility_bits[1] | visibility_bits[2]) == 0)
 		{
-			boost::optional<detail::TriangleToRasterize> t = detail::process_prospective_tri(clipspace_vertices[tri_indices[0]], clipspace_vertices[tri_indices[1]], clipspace_vertices[tri_indices[2]], viewport_width, viewport_height, enable_backface_culling);
+			std::optional<detail::TriangleToRasterize> t = detail::process_prospective_tri(clipspace_vertices[tri_indices[0]], clipspace_vertices[tri_indices[1]], clipspace_vertices[tri_indices[2]], viewport_width, viewport_height, enable_backface_culling);
 			if (t) {
 				triangles_to_raster.push_back(*t);
 			}
@@ -223,7 +222,7 @@ inline std::pair<core::Image4u, core::Image1d> render(core::Mesh mesh, glm::tmat
 		{
 			for (unsigned char k = 0; k < vertices.size() - 2; k++)
 			{
-				boost::optional<detail::TriangleToRasterize> t = detail::process_prospective_tri(vertices[0], vertices[1 + k], vertices[2 + k], viewport_width, viewport_height, enable_backface_culling);
+				std::optional<detail::TriangleToRasterize> t = detail::process_prospective_tri(vertices[0], vertices[1 + k], vertices[2 + k], viewport_width, viewport_height, enable_backface_culling);
 				if (t) {
 					triangles_to_raster.push_back(*t);
 				}
