@@ -24,17 +24,14 @@
 
 #include "glm/gtc/matrix_transform.hpp"
 
+#include "cxxopts.hpp"
+
 #include "opencv2/core/core.hpp"
 #include "opencv2/highgui/highgui.hpp"
-
-#include "boost/program_options.hpp"
-#include "boost/filesystem/path.hpp"
 
 #include <iostream>
 
 using namespace eos;
-namespace po = boost::program_options;
-namespace fs = boost::filesystem;
 using std::vector;
 using std::cout;
 using std::endl;
@@ -48,41 +45,38 @@ using std::endl;
  */
 int main(int argc, char *argv[])
 {
-	fs::path model_file, output_file;
+	std::string model_file, output_file;
 	vector<float> shape_coefficients, colour_coefficients;
 
 	try {
-		po::options_description desc("Allowed options");
-		desc.add_options()
-			("help",
-				"produce help message")
-			("model", po::value<fs::path>(&model_file)->required(),
-				"an eos .bin Morphable Model file")
-			("shape-coeffs", po::value<vector<float>>(&shape_coefficients)->multitoken(),
-				"optional parameter list of shape coefficients. All not specified will be set to zero. E.g.: '--shape-coeffs 0.0 1.5'. If omitted, the mean is used.")
-			("colour-coeffs", po::value<vector<float>>(&colour_coefficients)->multitoken(),
-				"optional parameter list of colour coefficients. All not specified will be set to zero. E.g.: '--colour-coeffs 0.0 1.5'. If omitted, the mean is used.")
-			("output", po::value<fs::path>(&output_file)->default_value("output.obj"),
-				"name of the output obj file (including .obj). Can be a full path.")
+		cxxopts::Options options("generate-obj");
+		options.add_options()
+			("help", "produce help message")
+			("model", "an eos .bin Morphable Model file",
+				cxxopts::value<std::string>(model_file))
+			("shape-coeffs", "optional parameter list of shape coefficients. All not specified will be set to zero. E.g.: '--shape-coeffs 0.0 1.5'. If omitted, the mean is used.",
+				cxxopts::value<vector<float>>(shape_coefficients)->multitoken()) // optional arg
+			("colour-coeffs", "optional parameter list of colour coefficients. All not specified will be set to zero. E.g.: '--colour-coeffs 0.0 1.5'. If omitted, the mean is used.",
+				cxxopts::value<vector<float>>(colour_coefficients)->multitoken()) // optional arg
+			("output", "name of the output obj file (including .obj). Can be a full path.",
+				cxxopts::value<std::string>(output_file)->default_value("output.obj"))
 			;
 
-		po::variables_map vm;
 		// disabling short options to allow negative values for the coefficients, e.g. '--shape-coeffs 0.0 -1.5'
-		po::store(po::parse_command_line(argc, argv, desc, po::command_line_style::unix_style ^ po::command_line_style::allow_short), vm);
-		if (vm.count("help")) {
-			cout << "Usage: generate-obj [options]" << endl;
-			cout << desc;
+		// Todo: Check how this works with cxxopts!
+		options.parse(argc, argv);
+		if (options.count("help")) {
+			cout << options.help() << endl;
 			return EXIT_SUCCESS;
 		}
-		po::notify(vm);
 	}
-	catch (const po::error& e) {
+	catch (const cxxopts::OptionException& e) {
 		cout << "Error while parsing command-line arguments: " << e.what() << endl;
 		cout << "Use --help to display a list of options." << endl;
 		return EXIT_FAILURE;
 	}
 
-	morphablemodel::MorphableModel morphable_model = morphablemodel::load_model(model_file.string());
+	morphablemodel::MorphableModel morphable_model = morphablemodel::load_model(model_file);
 
 	if (shape_coefficients.size() < morphable_model.get_shape_model().get_num_principal_components()) {
 		shape_coefficients.resize(morphable_model.get_shape_model().get_num_principal_components());
@@ -96,7 +90,7 @@ int main(int argc, char *argv[])
 
 	core::write_obj(sample_mesh, output_file.string());
 	core::Image4u rendering;
-	std::tie(rendering, std::ignore) = render::render(sample_mesh, glm::mat4x4(1.0f), glm::ortho(-130.0f, 130.0f, -130.0f, 130.0f), 512, 512, boost::none, true, false, false);
+	std::tie(rendering, std::ignore) = render::render(sample_mesh, glm::mat4x4(1.0f), glm::ortho(-130.0f, 130.0f, -130.0f, 130.0f), 512, 512, std::nullopt, true, false, false);
 	output_file.replace_extension(".png");
 	cv::imwrite(output_file.string(), core::to_mat(rendering));
 
