@@ -20,16 +20,13 @@
 #include "eos/morphablemodel/MorphableModel.hpp"
 #include "eos/morphablemodel/io/cvssp.hpp"
 
-#include "boost/program_options.hpp"
-#include "boost/filesystem.hpp"
+#include "cxxopts.hpp"
 
 #include <vector>
 #include <iostream>
 #include <fstream>
 
 using namespace eos;
-namespace po = boost::program_options;
-namespace fs = boost::filesystem;
 using std::cout;
 using std::endl;
 
@@ -39,32 +36,28 @@ using std::endl;
  */
 int main(int argc, char *argv[])
 {
-	fs::path scmmodelfile, isomapfile, outputfile;
+	std::string scmmodelfile, isomapfile, outputfile;
 	bool save_shape_only;
 	try {
-		po::options_description desc("Allowed options");
-		desc.add_options()
-			("help,h",
-				"display the help message")
-			("model,m", po::value<fs::path>(&scmmodelfile)->required(),
-				"a CVSSP .scm Morphable Model file")
-			("isomap,t", po::value<fs::path>(&isomapfile),
-				"optional isomap containing the texture mapping coordinates")
-			("shape-only,s", po::value<bool>(&save_shape_only)->default_value(false)->implicit_value(true),
-				"save only the shape-model part of the full 3DMM")
-			("output,o", po::value<fs::path>(&outputfile)->required()->default_value("converted_model.bin"),
-				"output filename for the Morphable Model in cereal binary format")
+		cxxopts::Options options("scm-to-cereal");
+		options.add_options()
+			("help,h", "display the help message")
+			("model,m", "a CVSSP .scm Morphable Model file",
+				cxxopts::value<std::string>(scmmodelfile))
+			("isomap,t", "optional isomap containing the texture mapping coordinates",
+				cxxopts::value<std::string>(isomapfile)) // optional arg
+			("shape-only,s", "save only the shape-model part of the full 3DMM",
+				cxxopts::value<bool>(save_shape_only)->default_value(false)->implicit_value(true)) // optional arg
+			("output,o", "output filename for the Morphable Model in cereal binary format",
+				cxxopts::value<std::string>(outputfile)->default_value("converted_model.bin"))
 			;
-		po::variables_map vm;
-		po::store(po::command_line_parser(argc, argv).options(desc).run(), vm);
-		if (vm.count("help")) {
-			cout << "Usage: scm-to-cereal [options]" << endl;
-			cout << desc;
+		options.parse(argc, argv);
+		if (options.count("help")) {
+			cout << options.help() << endl;
 			return EXIT_SUCCESS;
 		}
-		po::notify(vm);
 	}
-	catch (const po::error& e) {
+	catch (const cxxopts::OptionException& e) {
 		cout << "Error while parsing command-line arguments: " << e.what() << endl;
 		cout << "Use --help to display a list of options." << endl;
 		return EXIT_FAILURE;
@@ -77,12 +70,12 @@ int main(int argc, char *argv[])
 	{
 		// Save only the shape model - to generate the public sfm_shape_3448.bin
 		morphablemodel::MorphableModel shape_only_model(morphable_model.get_shape_model(), morphablemodel::PcaModel(), morphable_model.get_texture_coordinates());
-		morphablemodel::save_model(shape_only_model, outputfile.string());
+		morphablemodel::save_model(shape_only_model, outputfile);
 	}
 	else {
-		morphablemodel::save_model(morphable_model, outputfile.string());
+		morphablemodel::save_model(morphable_model, outputfile);
 	}
 
-	cout << "Saved converted model as " << outputfile.string() << "." << endl;
+	cout << "Saved converted model as " << outputfile << "." << endl;
 	return EXIT_SUCCESS;
 }
